Feat: Add DiskImageSource
